requirements.txt
numpy==1.19.3
scipy==1.5.4
matplotlib==3.3.3
scikit-image==0.18.1
natsort==7.1.1
opencv-python==4.5.3.56
datetime==4.3
streamlit==0.89.0
Pillow==8.3.2